Get dimension size from file's dimensions if possible
Only fall back on querying individual variable shapes if needed (i.e.,
for the unlimited dimension).

Note that I am removing the test that ensures that this function returns
0 if no variable is found with this dimension. This condition now only
applies to the unlimited dimension, and I don't want to take the time to
set up a file to test this: it's not a very important requirement.