Allow files under scipy to be exact duplicates of the scipy files
The point of this is so that we can more easily make updates: we can
just plop new versions in without having to make any changes.

This required some reorganization and adding an __init__.py file. The
latter renames cprnc_py.netcdf.scipy to just scipy, so it can be
accessed that way from netcdf.py. I'm not sure if this is the cleanest
or best way to accomplish this, but it seems to work.